requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
datasets>=2.18.0               # HuggingFace datasets — loads nlphuji/flickr30k
Pillow>=9.0.0
numpy>=1.24.0
opencv-contrib-python>=4.8.0   # for selective search proposals (optional, grid fallback exists)
gradio>=4.0.0                  # for demo